+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 
 ### Changed
 
+- Spotless predeclares dependencies
 - `root.project.name` property is not mandatory in Hubdle Settings
 
 ### Deprecated